finalize integration into GCode.cpp export functions, remove unused edge grids
